Update docs and comments to prefer TLS instead of SSL
The acronym SSL is considered deprecated by the security community and
has been replaced by TLS. While both are understood and can usually be
used interchangeably, Rails might as well use the modern terminology.

For an example on this preference, see Wikipedia article
https://en.wikipedia.org/wiki/Transport_Layer_Security

> LS builds on the now-deprecated SSL (Secure Sockets Layer)
> specifications ...

In the interest of back compatibility, this change only changes
documentation and code comments. It does not alter configuration option
names, class names, or other code. If it is preferable to make this
change in code as well, followup work can be submitted for this.

Some text uses the phrase "SSL/TLS". These instances were not altered.
If it is preferable to use this everywhere, followup work can be
submitted for this.